reorg of Haskell sources, better sensor model
The program ultimately computes motor velocity. This is computed from
desired flow, which is what the CMV program deals with. This should be a
nice factoring: ventilation modes can deal in desired flow, and there
can be one program which translates the desired flow into a motor
velocity.

The C simulation has been updated for the spec changes. Also, the UI
update is now triggerd by the spec, once every 10ms, and it is delivered
parameters.

The simulation itself is still stupid and naive, but we'll work on that
later. Next steps are to come up with a somewhat reasonable model in
which the flow sensor value is computed from the motor velocity control.